+++ src/opossum_lib/opossum/opossum_file_writer.py
