+++ tests/test_cli.py
@@ -12,8 +12,8 @@ from click.testing import CliRunner, Result
 
 from opossum_lib.cli import generate
 from opossum_lib.opossum.constants import INPUT_JSON_NAME, OUTPUT_JSON_NAME
-from opossum_lib.opossum.file_generation import OpossumFileWriter
 from opossum_lib.opossum.opossum_file import OpossumPackage
+from opossum_lib.opossum.opossum_file_writer import OpossumFileWriter
 from tests.test_setup.opossum_file_faker_setup import OpossumFileFaker
 
 test_data_path = Path(__file__).resolve().parent / "data"
