+++ tests/test_opossum/test_file_generation.py
@@ -6,8 +6,8 @@ from pathlib import Path
 from zipfile import ZipFile
 
 from opossum_lib.opossum.constants import INPUT_JSON_NAME, OUTPUT_JSON_NAME
-from opossum_lib.opossum.file_generation import OpossumFileWriter
 from opossum_lib.opossum.opossum_file_content import OpossumFileContent
+from opossum_lib.opossum.opossum_file_writer import OpossumFileWriter
 from tests.test_setup.opossum_file_faker_setup import OpossumFileFaker
